// src/app/models/admin/capphephoatdongkhoangsan/cpthamdokhoangsan.model.ts
export class OutputCpThamDoKhoangSanModel {
  idcapphepthamdo: string;
  idkhuvucthamdo: string;
  tenkhuvucthamdo: string;
  diadiem: string;
  dientichthamdo: number;
  dientichtralai: number;
  chieusauthamdotu: number;
  chieusauthamdoden: number;
  thoihanthamdo: number;
  ngaybdthamdo: string;
  ngayktthamdo: string;
  mucdichsudungkhoangsan: string;
  dangkhoangsan: string;
  donvidientich: string;
  donvithoihan: string;
  donvichieusau: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  idgiayphep: string;
  sogiayphep: string;
  siteid: string;
}

export class InputCpThamDoKhoangSanModel {
 idcapphepthamdo: string;
 tenkhuvucthamdo: string;
 diadiem: string;
 dientichthamdo: number;
 dientichtralai: number;
 chieusauthamdotu: number;
 chieusauthamdoden: number;
 thoihanthamdo: number;
 ngaybdthamdo: string;
 ngayktthamdo: string;
 mucdichsudungkhoangsan: string;
 dangkhoangsan: string;
 donvidientich: string;
 donvithoihan: string;
 donvichieusau: string;
 hequychieu: string;
 idgiayphep: string;
 siteid: string;
}